Fix task counter type in wcrt.c.

+++ src/wcrt.c
@@ -13,11 +13,11 @@ int main(int argc, char *argv[])
 
      while (scanf("{%ld,%ld,%ld}\n", &(tasks[n].r), &(tasks[n].c), &(tasks[n].p)) == 3) n++;
 
-     size_t j;
+     int j;
      for (j = n-1; j >= 0; j--)
      {
 	  rt_j = gap_enum(tasks, n, j, 1);
-	  printf("RT(τ[%ld] = {r=%ld,c=%ld,p=%ld}) = %ld\n",
+	  printf("RT(τ[%d] = {r=%ld,c=%ld,p=%ld}) = %ld\n",
 		 j, tasks[j].r, tasks[j].c, tasks[j].p, rt_j);
 	  if (rt_j < 0) break;
      }
